delete useless caching classes; structured project;
