+++ Ukranian-Culture.Backend/Services/ArticleTilesService.cs
@@ -3,8 +3,6 @@ using AutoMapper;
 using Contracts;
 using Entities.DTOs;
 using Entities.Models;
-using Microsoft.Extensions.Caching.Memory;
-using Repositories.CachingRepository;
 
 namespace Ukranian_Culture.Backend.Services;
 
@@ -58,22 +56,4 @@ public sealed class ArticleTilesService : IArticleTileService
         => await _repositoryManager
             .ArticleLocales
             .GetArticlesLocaleByConditionAsync(func, ChangesType.AsNoTracking);
-}
-
-public class CachingArticleTileService : IArticleTileService
-{
-    private readonly IArticleTileService _articleTileService;
-    private readonly CachingHelperService _cachingHelper;
-
-    public CachingArticleTileService(IArticleTileService articleTileService, IMemoryCache memoryCache)
-    {
-        _articleTileService = articleTileService;
-        _cachingHelper = new CachingHelperService("ArticleTile", memoryCache);
-    }
-
-    public Task<IEnumerable<ArticleTileDto>> TryGetArticleTileDto
-        (Guid cultureId, Expression<Func<Article, bool>> conditionToFindArticles)
-        => _cachingHelper.GetCachedResultAsync(
-            conditionToFindArticles.ToString(),
-            () => _articleTileService.TryGetArticleTileDto(cultureId, conditionToFindArticles)!)!;
 }
